requirements.txt
PyQt5==5.15.4
pandas
khayyam
python-docx
comtypes
PyPDF2
